+++ src/App.js
@@ -1,8 +1,7 @@
 import './App.css';
 import { ProjectCard } from './Components'
+import SignUp from './Components/SignUp/SignUp.component';
 import { fakeData } from './fakeData'
-import { Field, Form } from 'react-final-form'
-import axios from 'axios'
 
 //TODO
 //Header
@@ -12,43 +11,13 @@ import axios from 'axios'
 //Post a project form
 //
 
-const onSubmit = async values => {
-  console.log(values)
-  return await axios.post('http://localhost:8080/api/auth/local/signup', values).then(res => console.log(res)).catch(err => console.log(err))
-}
-// http://localhost:8080/api/projects
+
 
 const App = () => {
 
   return (
     <div className="App">
-
-      
-      <Form
-        onSubmit={onSubmit}
-        render={({ handleSubmit }) => (
-          <form onSubmit={handleSubmit}>
-            <div>
-              <div>Sign Up Form</div>
-              <label>Email</label>
-              <Field
-                name="email"
-                component="input"
-                type="email"
-                placeholder="email@youremail.com"
-              />
-              <label>Password</label>
-              <Field
-                name="password"
-                component="input"
-                type="password"
-              />
-            </div>
-            <button>Submit</button>
-          </form>
-        )}
-      />
-      <div>hola</div>
+      <SignUp/>
       {
         fakeData.map(element => {
           return (
