+++ src/Components/SignUp/SignUp.styles.css
@@ -1,0 +1,4 @@
+.form-container {
+    display: flex;
+    width: 50vw;
+}
